requirements.txt
requests~=2.30.0
python-dotenv~=1.0.0
google~=3.0.0
googleapiclient
tqdm